+++ src/infra/db/model/keyValue.go
@@ -1,8 +1,11 @@
 package tkInfraDbModel
 
+import "time"
+
 type KeyValue struct {
-	Key   string `gorm:"primaryKey"`
-	Value string `gorm:"not null"`
+	Key       string `gorm:"primaryKey"`
+	Value     string `gorm:"not null"`
+	ExpiresAt *time.Time
 }
 
 func (KeyValue) TableName() string {